ai/langchain-skills-agent/src/pyproject.toml
[project]
name = "langchain-skills-agent"
version = "1.0.0"
description = "A LangChain-powered agent with skill-based architecture for weather, travel, and currency conversion"
readme = "README.md"
requires-python = ">=3.13,<3.14.0"
dependencies = [
    "acp-sdk>=1.0.3",
    "fastapi>=0.135.0",
    "langchain>=1.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=1.2.0",
    "langchain-ibm>=1.0.3",
    "python-dotenv>=1.2.0",
    "pyyaml>=6.0.0",
    "uvicorn==0.34.0",
]

[tool.poetry]
package-mode = false

[dependency-groups]
dev = ["ruff>=0.11.10"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
